requirements.txt
matplotlib
numpy
tensorflow
tensorboard
pandas
Pillow